+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,6 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".TeamFinderApp"
         android:theme="@style/Theme.TeamFinder"
         tools:targetApi="31">
         <activity
