--- app/src/main/java/com/teamfinder/teamfinder/TeamFinderApp.kt
@@ -1,17 +1,0 @@
-package com.teamfinder.teamfinder
-
-import android.app.Application
-import com.teamfinder.teamfinder.di.AppComponent
-import com.teamfinder.teamfinder.di.DaggerAppComponent
-
-class TeamFinderApp : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        component = DaggerAppComponent.create()
-    }
-
-    companion object {
-        lateinit var component: AppComponent
-    }
-}
